+++ src/data-workspace/category-combo-table-body-pivoted/category-combo-table-body-pivoted.jsx
@@ -5,6 +5,7 @@ import React from 'react'
 import { useMetadata, selectors } from '../../shared/index.js'
 import { DataEntryCell, DataEntryField } from '../data-entry-cell/index.js'
 import { getFieldId } from '../get-field-id.js'
+import { TableBodyHiddenByFiltersRow } from '../table-body-hidden-by-filter-row.js'
 import styles from '../table-body.module.css'
 import { generateFormMatrix } from './generate-form-matrix/index.js'
 
@@ -24,9 +25,9 @@ export const PivotedCategoryComboTableBody = React.memo(
         categoryCombo,
         dataElements,
         greyedFields,
-        /*
         filterText,
         globalFilterText,
+        /*
         maxColumnsInSection,
         renderRowTotals,
         renderColumnTotals,*/
@@ -55,15 +56,49 @@ export const PivotedCategoryComboTableBody = React.memo(
             })
             .flat()
 
+        const filterCOCBySearch = (elementName) => {
+            const search =
+                displayOptions.pivotMode !== 'move_categories' &&
+                filterText.toLowerCase()
+            return search ? elementName.toLowerCase().includes(search) : true
+        }
+
+        const filterDataElementsBySearch = (elementName) => {
+            // local search should always work on rows - so disable localSearch when transposed (filterCOCBySearch would do the filter)
+            const localSearch =
+                displayOptions.pivotMode === 'move_categories' &&
+                filterText.toLowerCase()
+            const globalSearch = globalFilterText.toLowerCase()
+            return localSearch || globalSearch
+                ? (localSearch &&
+                      elementName.toLowerCase().includes(localSearch)) ||
+                      (globalSearch &&
+                          elementName.toLowerCase().includes(globalSearch))
+                : true
+        }
+
+        const filteredDataElements = dataElements.filter((de) =>
+            filterDataElementsBySearch(de.displayFormName)
+        )
+        const filteredSortedCOCS = sortedCOCs.filter((coc) =>
+            filterCOCBySearch(coc.displayName)
+        )
+
         const options = {
             metadata,
             categoryOptionsDetails,
-            sortedCOCs,
+            sortedCOCs: filteredSortedCOCS,
+            totalRows: sortedCOCs.length,
             categories,
-            dataElements,
+            dataElements: filteredDataElements,
         }
 
         const rowsMatrix = generateFormMatrix(options, displayOptions)
+
+        const hiddenItemsCount =
+            displayOptions.pivotMode === 'move_categories'
+                ? dataElements.length - filteredDataElements.length
+                : sortedCOCs.length - filteredSortedCOCS.length
 
         return (
             <>
@@ -147,6 +182,11 @@ export const PivotedCategoryComboTableBody = React.memo(
                         </TableRow>
                     )
                 })}
+                {hiddenItemsCount > 0 && (
+                    <TableBodyHiddenByFiltersRow
+                        hiddenItemsCount={hiddenItemsCount}
+                    />
+                )}
             </>
         )
     }
@@ -173,6 +213,8 @@ PivotedCategoryComboTableBody.propTypes = {
         pivotMode: PropTypes.oneOf(['move_categories', 'pivot']),
         pivotedCategory: PropTypes.string,
     }),
+    filterText: PropTypes.string,
     /** Greyed fields is a Set where .has(fieldId) is true if that field is greyed/disabled */
+    globalFilterText: PropTypes.string,
     greyedFields: PropTypes.instanceOf(Set),
 }