+++ src/data-workspace/custom-form/custom-form-total-cell.jsx
@@ -1,13 +1,15 @@
 import PropTypes from 'prop-types'
 import React, { useState, useEffect } from 'react'
-import { useForm } from 'react-final-form'
-import { useBlurredField } from '../../shared/use-blurred-field.js'
+import {
+    useBlurredField,
+    useValueStore,
+    useEntryFormStore,
+} from '../../shared/index.js'
 import { TotalCell } from '../category-combo-table-body/total-cells.jsx'
-import { getFieldId, parseFieldId } from '../get-field-id.js'
+import { parseFieldId } from '../get-field-id.jsx'
 
-const computeTotal = (dataElementId, formState) => {
-    const { values, hasValidationErrors, errors } = formState
-    const dataElementValues = values[dataElementId]
+const computeTotal = (dataElementId, dataValues, formErrors) => {
+    const dataElementValues = dataValues?.[dataElementId] || {}
 
     if (!dataElementValues) {
         return null
@@ -15,31 +17,37 @@ const computeTotal = (dataElementId, formState) => {
 
     // Initialise sum as null and only start counting when numerical values
     // are encountered to avoid rendering zeros when the sum isn't actually zero
-    return Object.entries(dataElementValues).reduce((sum, [cocId, value]) => {
-        const fieldHasError =
-            hasValidationErrors && errors[getFieldId(dataElementId, cocId)]
+    return Object.entries(dataElementValues).reduce(
+        (sum, [cocId, valueDetails]) => {
+            const fieldHasError = formErrors?.[dataElementId]?.[cocId]
 
-        if (!fieldHasError && !isNaN(value)) {
-            sum = isNaN(sum) ? value : sum + Number(value)
-        }
-        return sum
-    }, null)
+            const value = valueDetails?.value
+
+            if (!fieldHasError && !isNaN(value)) {
+                sum = isNaN(sum) ? value : sum + Number(value)
+            }
+            return sum
+        },
+        null
+    )
 }
 
 export const CustomFormTotalCell = ({ dataElementId }) => {
-    const form = useForm()
+    const dataValues = useValueStore((state) => state.getDataValues())
+    const formErrors = useEntryFormStore((state) => state.getErrors())
+
     const blurredField = useBlurredField()
     const [total, setTotal] = useState(() =>
-        computeTotal(dataElementId, form.getState())
+        computeTotal(dataElementId, dataValues, formErrors)
     )
 
     useEffect(() => {
         const { dataElementId: blurredFieldDataElementId } =
             parseFieldId(blurredField)
         if (blurredFieldDataElementId === dataElementId) {
-            setTotal(computeTotal(dataElementId, form.getState()))
+            setTotal(computeTotal(dataElementId, dataValues, formErrors))
         }
-    }, [blurredField, dataElementId, form])
+    }, [blurredField, dataElementId, dataValues, formErrors])
 
     return <TotalCell>{total}</TotalCell>
 }