+++ src/data-workspace/indicators-table-body/compute-indicator-value.js
@@ -1,7 +1,8 @@
 import { Parser } from 'expr-eval'
-import { getIn } from 'final-form'
 import { parseFieldId as parseFieldOperand } from '../get-field-id.js'
 
+export const NONCALCULABLE_VALUE = 'noncalculable_value'
+export const MATHEMATICALLY_INVALID_VALUE = 'mathematically_invalid_value'
 /**
  * --- INDICATOR VALUE CALCULATION ---
  * The general formula for computing an indicator value is:
@@ -18,7 +19,7 @@ const evaluate = (expression) => {
     try {
         return parser.parse(expression).evaluate()
     } catch {
-        return ''
+        return NONCALCULABLE_VALUE
     }
 }
 /*
@@ -38,9 +39,9 @@ const getDataElementTotalValue = (dataElementCocValues) => {
         return 0
     }
 
-    return Object.values(dataElementCocValues).reduce((sum, value) => {
-        if (!isNaN(value)) {
-            sum = sum + Number(value)
+    return Object.values(dataElementCocValues).reduce((sum, valueObject) => {
+        if (!isNaN(valueObject?.value)) {
+            sum = sum + Number(valueObject?.value)
         }
         return sum
     }, 0)
@@ -69,7 +70,7 @@ const getDataElementTotalValue = (dataElementCocValues) => {
  * @param {Object} values ReactFinalForm `values`
  * @returns {string} a parsed expression template
  */
-const parseExpressionTemplate = (expression, values) => {
+const parseExpressionTemplate = (expression, dataValues) => {
     const matches = expression.match(operandInterpolationPattern)
 
     /*
@@ -83,13 +84,14 @@ const parseExpressionTemplate = (expression, values) => {
 
     return matches.reduce((substitudedExpression, match) => {
         const operand = match.replace(/[#{}]/g, '')
-        const { categoryOptionComboId } = parseFieldOperand(operand)
+        const { dataElementId, categoryOptionComboId } =
+            parseFieldOperand(operand)
         const value =
             (categoryOptionComboId
                 ? // For COCs we read the data directly from the form values
-                  getIn(values, operand)
+                  dataValues?.[dataElementId]?.[categoryOptionComboId]?.value
                 : // For data elements we need the sum of all COC form values
-                  getDataElementTotalValue(values[operand])) || 0
+                  getDataElementTotalValue(dataValues?.[dataElementId])) || 0
 
         return substitudedExpression.replace(match, value)
     }, expression)
@@ -119,7 +121,7 @@ export const round = (value, decimals) => {
  * @param {string} options.denominator Indicator expression template
  * @param {string} options.numerator Indicator expression template
  * @param {number} options.factor Indicator multiplier
- * @param {{values: object}} options.formState ReactFinalForm formState
+ * @param {Object} options.dataValues {[de]:{[coc]:{...rest,value}}}
  * @returns {number} Indicator value
  */
 
@@ -127,23 +129,40 @@ export const computeIndicatorValue = ({
     denominator,
     numerator,
     factor,
-    formState,
+    dataValues,
     decimals,
 }) => {
-    const numeratorExpression = parseExpressionTemplate(
-        numerator,
-        formState.values
-    )
+    const numeratorExpression = parseExpressionTemplate(numerator, dataValues)
     const denominatorExpression = parseExpressionTemplate(
         denominator,
-        formState.values
+        dataValues
     )
     const numeratorValue = evaluate(numeratorExpression)
     const denominatorValue = evaluate(denominatorExpression)
+
+    if (
+        numeratorValue === NONCALCULABLE_VALUE ||
+        denominatorValue === NONCALCULABLE_VALUE
+    ) {
+        return {
+            value: NONCALCULABLE_VALUE,
+            numeratorValue,
+            denominatorValue,
+        }
+    }
     const indicatorValue = (numeratorValue / denominatorValue) * factor
     const isReadableNumber = isFinite(indicatorValue) && !isNaN(indicatorValue)
+
     if (!isReadableNumber) {
-        return ''
+        return {
+            value: MATHEMATICALLY_INVALID_VALUE,
+            numeratorValue,
+            denominatorValue,
+        }
     }
-    return decimals ? round(indicatorValue, decimals) : indicatorValue
+    return {
+        value: decimals ? round(indicatorValue, decimals) : indicatorValue,
+        numeratorValue,
+        denominatorValue,
+    }
 }