+++ src/data-workspace/indicators-table-body/compute-indicator-value.test.js
@@ -1,81 +1,93 @@
 import { computeIndicatorValue, round } from './compute-indicator-value.js'
 
 describe('computeIndicatorValue', () => {
-    const formState = {
-        values: {
-            a: {
-                a1: '1',
-                a2: '2',
-            },
-            b: {
-                b1: '2',
-            },
-            c: {
-                c1: '3',
-            },
-            d: {
-                d1: '1',
-                d2: '1',
-                d3: '1',
-                d4: '1',
-                d5: '1',
-            },
+    const dataValues = {
+        a: {
+            a1: { value: '1' },
+            a2: { value: '2' },
+        },
+        b: {
+            b1: { value: '2' },
+        },
+        c: {
+            c1: { value: '3' },
+        },
+        d: {
+            d1: { value: '1' },
+            d2: { value: '1' },
+            d3: { value: '1' },
+            d4: { value: '1' },
+            d5: { value: '1' },
         },
     }
     it('produces a correct value for category option combos', () => {
-        const value = computeIndicatorValue({
+        const result = computeIndicatorValue({
             numerator: '#{a.a1}+#{b.b1}*#{c.c1}', // 1+2*3=7
             denominator: '#{d.d1}*#{c.c1}', // 1*3=3
             factor: 1,
-            formState,
+            dataValues,
         })
-        expect(value).toBe(7 / 3)
+        expect(result.value).toBe(7 / 3)
+        expect(result.numeratorValue).toBe(7)
+        expect(result.denominatorValue).toBe(3)
     })
     it('produces a correct value for data elements', () => {
-        const value = computeIndicatorValue({
+        const result = computeIndicatorValue({
             numerator: '#{a}+#{b.b1}*#{c.c1}', // 3+2*3=9
             denominator: '#{d}*#{c}', // 5*3=15
             factor: 1,
-            formState,
+            dataValues,
         })
-        expect(value).toBe(9 / 15)
+        expect(result.value).toBe(9 / 15)
+        expect(result.numeratorValue).toBe(9)
+        expect(result.denominatorValue).toBe(15)
     })
     it('produces a correct value when a factor is present', () => {
-        const value = computeIndicatorValue({
+        const result = computeIndicatorValue({
             numerator: '#{a}+#{b.b1}*#{c.c1}', // 3+2*3=9
             denominator: '#{d}*#{c}', // 5*3=15
             factor: 100,
-            formState,
+            dataValues,
         })
-        expect(value).toBe((9 / 15) * 100)
+        expect(result.value).toBe((9 / 15) * 100)
+        expect(result.numeratorValue).toBe(9)
+        expect(result.denominatorValue).toBe(15)
     })
-    it('returns an empty string when computed value would be NaN', () => {
-        const value = computeIndicatorValue({
-            numerator: '#{nothing}',
+    it('returns noncalculable_value when expressions cannot be parsed', () => {
+        const result = computeIndicatorValue({
+            numerator: '#{a}.periodOffset(-1)',
             denominator: '#{nothing_also}',
             factor: 1,
-            formState,
+            dataValues,
         })
-        expect(value).toBe('')
+        expect(result.value).toBe('noncalculable_value')
     })
-    it('returns an empty string when computed value would be Infinity', () => {
-        const value = computeIndicatorValue({
+    it('returns mathematically_invalid_value when computed value would be Infinity', () => {
+        const result = computeIndicatorValue({
             numerator: '#{a}+#{b.b1}*#{c.c1}', // 3+2*3=9
             denominator: '#{nothing}',
             factor: 1,
-            formState,
+            dataValues,
         })
-        expect(value).toBe('')
+        expect(result.value).toBe('mathematically_invalid_value')
+    })
+    it('returns mathematically_invalid_value when computed value would be NaN', () => {
+        const result = computeIndicatorValue({
+            numerator: 'sqrt(-1)',
+            denominator: '#{nothing}',
+            factor: 1,
+        })
+        expect(result.value).toBe('mathematically_invalid_value')
     })
     it('returns an appropriately rounded value', () => {
-        const value = computeIndicatorValue({
+        const result = computeIndicatorValue({
             numerator: '#{a.a1}', // 1
             denominator: '#{c.c1}', // 3
             factor: 100,
-            formState,
+            dataValues,
             decimals: 3,
         })
-        expect(value).toBe(33.333)
+        expect(result.value).toBe(33.333)
     })
 })
 
@@ -102,7 +114,7 @@ describe('round', () => {
         const decimals = 3
         const originalValue = NaN
         const roundedValue = round(originalValue, decimals)
-        expect(roundedValue).toBe(originalValue)
+        expect(roundedValue).toBe(NaN)
     })
     it('returns the original value if decimals is not provided', () => {
         const decimals = undefined